:truck: Move plugins directory content up
