+++ packages/core/src/batch.spec.ts
@@ -1,4 +1,4 @@
-import { Cuillere, batched, batchPlugin, callable, cuillere } from '..'
+import { Cuillere, batched, batchPlugin, callable, cuillere } from '.'
 
 const delay = (timeout: number) => new Promise(resolve => setTimeout(resolve, timeout))
