+++ packages/core/src/index.ts
@@ -1,8 +1,10 @@
+export * from './batch'
+export * from './concurrent'
 export * from './cuillere'
 export * from './effect'
 export * from './executable-promise'
 export * from './generator'
-export * from './plugins'
+export * from './plugin'
 export * from './operation'
 export { Task } from './stack'
 export * from './time'
