+++ packages/core/src/plugin.spec.ts
@@ -1,4 +1,4 @@
-import { Cuillere, Plugin, cuillere, next } from '..'
+import { Cuillere, Plugin, cuillere, next } from '.'
 
 const nodeMajorVersion = Number(process.versions.node.split('.')[0])
