--- packages/core/src/plugins/index.ts
@@ -1,3 +1,0 @@
-export * from './plugin'
-export * from './concurrent'
-export * from './batch'
